--- src/css/MultiStepBookingForm.css
@@ -1,354 +1,0 @@
-/* Multi-Step Dialog */
-.multi-step-dialog .MuiDialog-paper {
-  border-radius: 20px !important;
-  box-shadow: 0 25px 50px rgba(0, 0, 0, 0.25) !important;
-  overflow: hidden;
-  max-width: 1000px !important;
-  max-height: 95vh !important;
-}
-
-/* Header with Stepper */
-.multi-step-header {
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  color: white;
-  padding: 25px 30px 20px;
-  position: relative;
-  overflow: hidden;
-}
-
-.multi-step-header::before {
-  content: '';
-  position: absolute;
-  top: -50%;
-  right: -50%;
-  width: 100%;
-  height: 200%;
-  background: rgba(255, 255, 255, 0.1);
-  transform: rotate(45deg);
-}
-
-.close-button {
-  position: absolute !important;
-  top: 20px !important;
-  right: 25px !important;
-  background: rgba(255, 255, 255, 0.2) !important;
-  color: white !important;
-  width: 35px !important;
-  height: 35px !important;
-  border-radius: 50% !important;
-  font-size: 18px !important;
-  z-index: 3;
-  transition: all 0.3s ease !important;
-}
-
-.close-button:hover {
-  background: rgba(255, 255, 255, 0.3) !important;
-  transform: rotate(90deg);
-}
-
-/* Custom Stepper */
-.stepper-container {
-  position: relative;
-  z-index: 2;
-}
-
-.custom-stepper {
-  background: transparent !important;
-}
-
-.custom-stepper .MuiStepConnector-root {
-  top: 22px !important;
-  left: calc(-50% + 20px) !important;
-  right: calc(50% + 20px) !important;
-}
-
-.custom-stepper .MuiStepConnector-line {
-  height: 3px !important;
-  border: none !important;
-  background: rgba(255, 255, 255, 0.3) !important;
-  border-radius: 2px !important;
-}
-
-.custom-stepper .MuiStepConnector-root.Mui-active .MuiStepConnector-line,
-.custom-stepper .MuiStepConnector-root.Mui-completed .MuiStepConnector-line {
-  background: white !important;
-}
-
-.custom-step {
-  cursor: pointer !important;
-  transition: all 0.3s ease !important;
-}
-
-.custom-step:hover {
-  transform: translateY(-2px);
-}
-
-.step-icon {
-  width: 44px !important;
-  height: 44px !important;
-  border-radius: 50% !important;
-  background: rgba(255, 255, 255, 0.2) !important;
-  display: flex !important;
-  align-items: center !important;
-  justify-content: center !important;
-  color: white !important;
-  transition: all 0.3s ease !important;
-  backdrop-filter: blur(10px) !important;
-  border: 2px solid rgba(255, 255, 255, 0.3) !important;
-}
-
-.step-icon.active {
-  background: white !important;
-  color: #667eea !important;
-  border-color: white !important;
-  box-shadow: 0 4px 20px rgba(255, 255, 255, 0.3) !important;
-  transform: scale(1.1) !important;
-}
-
-.step-icon.completed {
-  background: rgba(255, 255, 255, 0.9) !important;
-  color: #667eea !important;
-  border-color: white !important;
-}
-
-.custom-step-label .MuiStepLabel-label {
-  color: rgba(255, 255, 255, 0.8) !important;
-  font-weight: 500 !important;
-  margin-top: 8px !important;
-  font-size: 14px !important;
-}
-
-.custom-step-label.Mui-active .MuiStepLabel-label {
-  color: white !important;
-  font-weight: 600 !important;
-}
-
-.custom-step-label.Mui-completed .MuiStepLabel-label {
-  color: rgba(255, 255, 255, 0.9) !important;
-}
-
-/* Content Area */
-.multi-step-content {
-  padding: 35px 30px !important;
-  background: linear-gradient(135deg, #f8fafc 0%, #e2e8f0 100%) !important;
-  max-height: calc(95vh - 250px) !important;
-  overflow-y: auto !important;
-}
-
-.step-content {
-  animation: slideInUp 0.4s ease-out;
-}
-
-@keyframes slideInUp {
-  from {
-    opacity: 0;
-    transform: translateY(20px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-/* Booking Form Container (Step 1) */
-.booking-form-container {
-  display: flex;
-  gap: 30px;
-  align-items: flex-start;
-}
-
-.booking-column {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-}
-
-/* Customer Form Container (Step 2) */
-.customer-form-container {
-  max-width: 800px;
-  margin: 0 auto;
-}
-
-.customer-form-grid {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
-
-/* Deposit Container (Step 3) */
-.deposit-container {
-  max-width: 600px;
-  margin: 0 auto;
-}
-
-.deposit-summary {
-  padding: 20px;
-  background: white;
-  border-radius: 16px;
-  border: 1px solid rgba(102, 126, 234, 0.2);
-}
-
-.deposit-summary h4 {
-  margin: 0 0 20px 0;
-  color: #2d3748;
-  font-size: 18px;
-  font-weight: 600;
-  text-align: center;
-  padding-bottom: 15px;
-  border-bottom: 2px solid #e2e8f0;
-}
-
-.summary-item {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 12px 0;
-  border-bottom: 1px solid #f1f5f9;
-}
-
-.summary-item:last-child {
-  border-bottom: none;
-  padding-top: 15px;
-  margin-top: 10px;
-  border-top: 2px solid #e2e8f0;
-}
-
-.summary-item span:first-child {
-  color: #64748b;
-  font-weight: 500;
-}
-
-.summary-item span:last-child {
-  color: #2d3748;
-  font-weight: 600;
-}
-
-.total-amount {
-  color: #667eea !important;
-  font-size: 18px !important;
-  font-weight: 700 !important;
-}
-
-/* Section Headers */
-.section-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 20px;
-}
-
-.section-header h3 {
-  font-size: 18px;
-  font-weight: 600;
-  color: #2d3748;
-  margin: 0;
-  display: flex;
-  align-items: center;
-}
-
-.section-header h3::before {
-  content: '';
-  width: 4px;
-  height: 20px;
-  background: linear-gradient(135deg, #667eea, #764ba2);
-  border-radius: 2px;
-  margin-right: 10px;
-}
-
-/* Booking Sections */
-.booking-section {
-  background: white;
-  border-radius: 16px;
-  padding: 25px;
-  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.08);
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  backdrop-filter: blur(10px);
-}
-
-.room-details-section {
-  padding: 20px;
-}
-
-/* Form Styling */
-.form-row {
-  margin-bottom: 20px;
-}
-
-.form-row-group {
-  display: flex;
-  gap: 15px;
-  margin-bottom: 20px;
-}
-
-.form-row-group .booking-textfield {
-  flex: 1;
-}
-
-/* Text Fields */
-.booking-textfield .MuiOutlinedInput-root {
-  border-radius: 12px !important;
-  background: #f8fafc !important;
-  transition: all 0.3s ease !important;
-  border: 2px solid #e2e8f0 !important;
-}
-
-.booking-textfield .MuiOutlinedInput-root:hover {
-  background: white !important;
-  transform: translateY(-1px);
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-}
-
-.booking-textfield .MuiOutlinedInput-root.Mui-focused {
-  background: white !important;
-  border-color: #667eea !important;
-  box-shadow: 0 0 0 3px rgba(102, 126, 234, 0.1) !important;
-  transform: translateY(-1px);
-}
-
-.booking-textfield .MuiOutlinedInput-notchedOutline {
-  border: none !important;
-}
-
-.booking-textfield .MuiInputLabel-root {
-  color: #64748b !important;
-  font-weight: 500 !important;
-}
-
-.booking-textfield .MuiInputLabel-root.Mui-focused {
-  color: #667eea !important;
-}
-
-.booking-textfield .MuiOutlinedInput-input {
-  padding: 12px 16px !important;
-  font-size: 15px !important;
-}
-
-.currency {
-  color: #64748b;
-  font-size: 14px;
-  font-weight: 500;
-  margin-left: 8px;
-}
-
-/* Room Detail Cards */
-.room-detail-card {
-  background: linear-gradient(135deg, #f8fafc, #e2e8f0);
-  border-radius: 12px;
-  padding: 20px;
-  margin-bottom: 15px;
-  border: 1px solid rgba(255, 255, 255, 0.5);
-  transition: all 0.3s ease;
-}
-
-.room-detail-card:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 8px 25px rgba(0, 0, 0, 0.1);
-}
-
-.room-detail-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 15px;
-}
-
