+++ src/css/OrderRoom.css
@@ -60,7 +60,7 @@
 }
 
 .order-room-filter-tat_ca {
-  background: linear-gradient(135deg, #e8f2e6, #d7b8c0, #f9e8d6);
+  background: linear-gradient(135deg, #607D8B, #455A64);
 }
 
 .order-room-filter-tat_ca:hover {
@@ -74,7 +74,7 @@
 }
 
 .order-room-filter-trong {
-  background: linear-gradient(135deg, #e0f2e1, #3abb3a);
+  background: linear-gradient(135deg, #4CAF50, #2E7D32);
 }
 
 .order-room-filter-trong:hover {
@@ -88,7 +88,7 @@
 }
 
 .order-room-filter-da_dat {
-  background: linear-gradient(135deg, #ffebee, #e74a5a);
+  background: linear-gradient(135deg, #3c65be, #3c65be);
 }
 
 .order-room-filter-da_dat:hover {
@@ -102,7 +102,7 @@
 }
 
 .order-room-filter-dang_sua {
-  background: linear-gradient(135deg, #fff8e1, #f1c94f);
+  background: linear-gradient(135deg, #FF9800, #F57C00);
 }
 
 .order-room-filter-dang_sua:hover {
@@ -117,130 +117,220 @@
 
 .order-room-grid {
   display: grid;
-  grid-template-columns: repeat(5, 1fr);
+  grid-template-columns: repeat(auto-fill, minmax(350px, 1fr));
   gap: 20px;
   margin-bottom: 20px;
 }
 
+/* Card Layout - Split design */
 .order-room-card {
   position: relative;
-  border-radius: 15px;
-  padding: 15px;
-  box-shadow: 0 8px 25px rgba(0, 0, 0, 0.1);
+  border-radius: 10px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
   background: white;
   transition: all 0.3s ease;
   overflow: hidden;
-  border-left: 6px solid transparent;
-  min-height: 167px;
+  min-height: 120px;
   display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  border: 1px solid rgba(0, 0, 0, 0.05);
   cursor: pointer;
+  border: 1px solid rgba(0, 0, 0, 0.05);
 }
 
 .order-room-card:hover {
-  transform: translateY(-5px);
-  box-shadow: 0 12px 35px rgba(0, 0, 0, 0.2);
+  transform: translateY(-3px);
+  box-shadow: 0 8px 25px rgba(0, 0, 0, 0.15);
 }
 
-.order-room-status-nhan_phong {
-  border-left-color: #F44336;
-  background: linear-gradient(135deg, #ffebee, #ffcdd2);
-}
-.order-room-status-trong {
-  border-left-color: #43A047;
-  background: linear-gradient(135deg, #e0f2e1, #b7d8b7);
-}
-.order-room-status-qua_han {
-  border-left-color: #FF9800;
-  background: linear-gradient(135deg, #ffe0b2, #ffcc80);
-}
-.order-room-status-da_dat {
-  border-left-color: #F44336;
-  background: linear-gradient(135deg, #ffebee, #ffcdd2);
-}
-.order-room-status-dang_sua {
-  border-left-color: #FFC107;
-  background: linear-gradient(135deg, #fff8e1, #ffecb3);
+/* Left Section - Dark colors */
+.order-room-left-section {
+  width: 120px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 15px;
+  position: relative;
 }
 
-.order-room-id {
-  font-size: 22px;
-  font-weight: 700;
-  color: #2c3e50;
-  background: rgba(255, 255, 255, 0.9);
-  padding: 8px 12px;
-  border-radius: 8px;
-  margin-bottom: 10px;
-  display: inline-block;
+/* Right Section - Light colors */
+.order-room-right-section {
+  flex: 1;
+  padding: 15px;
+  display: flex;
+  flex-direction: column;
+  position: relative;
+}
+
+/* Status-specific colors for right section */
+.order-room-status-trong .order-room-left-section {
+  background: #4CAF50; /* Màu xanh lá cho Trống */
+}
+
+.order-room-status-da_dat .order-room-left-section {
+  background: #3c65be; /* Màu đỏ cho Đã đặt */
+}
+
+.order-room-status-dang_sua .order-room-left-section {
+  background: #F57C00; /* Màu cam cho Đang sửa */
+}
+
+/* Status-specific colors for right section */
+.order-room-status-trong .order-room-right-section {
+  background: #E8F5E9; /* Xanh lá nhạt */
+}
+
+.order-room-status-da_dat .order-room-right-section {
+  background: #eeebff; /* Đỏ nhạt */
+}
+
+.order-room-status-dang_sua .order-room-right-section {
+  background: #FFF3E0; /* Cam nhạt */
+}
+
+/* Left section elements */
+.order-room-status-label {
+  font-size: 11px;
+  font-weight: 600;
   text-transform: uppercase;
-  border: 1px solid rgba(0, 0, 0, 0.1);
+  letter-spacing: 1px;
+  margin-bottom: 5px;
+  opacity: 0.9;
+}
+
+.order-room-number {
+  font-size: 28px;
+  font-weight: 700;
+  margin-bottom: 8px;
+  text-align: center;
+}
+
+.order-room-icon {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  opacity: 0.8;
+}
+
+.order-room-icon svg {
+  font-size: 20px;
+}
+
+/* Right section elements */
+.order-room-view-button {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  color: #666;
+  background: rgba(255, 255, 255, 0.5);
+  border-radius: 50%;
+  padding: 6px;
+  box-shadow: 0 2px 8px rgba(214, 15, 15, 0.1);
+  border: none;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 36px;
+  height: 36px;
+}
+
+.order-room-view-button:hover {
+  color: #333;
+  transform: scale(1.1);
+  background: rgba(255, 255, 255, 0.95);
+}
+
+.order-room-view-button svg {
+  width: 24px;
+  height: 24px;
+}
+
+.order-room-content {
+  margin-top: 25px;
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.order-room-status-text {
+  font-size: 18px;
+  font-weight: 600;
+  color: #2c3e50;
+  text-align: center;
+  margin-bottom: 5px;
+}
+
+.order-room-date-time {
+  font-size: 14px;
+  color: #555;
+  margin-bottom: 8px;
+  font-weight: 500;
 }
 
 .order-room-guest-name {
-  font-size: 15px;
-  color: #34495e;
-  font-weight: 500;
-  margin-bottom: 6px;
-  text-align: center;
-  padding: 4px;
-  background: rgba(255, 255, 255, 0.8);
-  border-radius: 5px;
-}
-
-.order-room-date {
-  font-size: 13px;
-  color: #7f8c8d;
+  font-size: 16px;
+  color: #2c3e50;
+  font-weight: 600;
   margin-bottom: 8px;
-  text-align: center;
-  padding: 4px;
-  background: rgba(255, 255, 255, 0.8);
-  border-radius: 5px;
 }
 
 .order-room-countdown {
-  font-size: 13px;
+  font-size: 14px;
   color: #e74c3c;
   font-weight: 600;
-  text-align: center;
-  padding: 6px;
-  background: rgba(231, 76, 60, 0.15);
-  border-radius: 5px;
-  margin-top: 8px;
-  border: 1px solid rgba(231, 76, 60, 0.3);
-}
-
-.order-room-guest {
-  font-size: 15px;
-  color: #2c3e50;
-  font-weight: 600;
-  text-align: center;
-  padding: 6px;
-  background: rgba(255, 255, 255, 0.8);
-  border-radius: 5px;
-  margin-top: auto;
-}
-
-.order-room-action-view {
-  color: #3498db;
-  transition: all 0.3s ease;
-  background: rgba(255, 255, 255, 0.9);
-  border-radius: 50%;
-  padding: 5px;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-}
-
-.order-room-action-view:hover {
-  color: #2980b9;
-  transform: scale(1.1);
-  background: rgba(52, 152, 219, 0.1);
+  font-family: 'Courier New', monospace;
+  background: rgba(231, 76, 60, 0.1);
+  padding: 4px 8px;
+  border-radius: 4px;
+  display: inline-block;
+  border: 1px solid rgba(231, 76, 60, 0.2);
 }
 
 .order-room-no-data {
   text-align: center;
-  padding: 20px;
+  padding: 40px 20px;
   color: #7f8c8d;
   font-size: 18px;
   font-weight: 500;
+  background: white;
+  border-radius: 10px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+}
+
+/* Responsive design */
+@media (max-width: 1400px) {
+  .order-room-grid {
+    grid-template-columns: repeat(auto-fill, minmax(320px, 1fr));
+  }
+}
+
+@media (max-width: 1200px) {
+  .order-room-grid {
+    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+  }
+}
+
+@media (max-width: 768px) {
+  .order-room-grid {
+    grid-template-columns: 1fr;
+  }
+  
+  .order-room-card {
+    min-height: 100px;
+  }
+  
+  .order-room-left-section {
+    width: 100px;
+    padding: 10px;
+  }
+  
+  .order-room-number {
+    font-size: 24px;
+  }
+  
+  .order-room-right-section {
+    padding: 10px;
+  }
 }
